agentic_rag_pipeline/requirements.txt
fastapi>=0.115.0
google-genai>=1.0.0
pinecone>=7.0.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
